feat(wm): add option to either minimize or hide
This commit adds a command to let the user decide if they want windows
to be hidden with SW_HIDE or minimized with SW_MINIMIZE when workspaces
are changed or window container stacks are cycled.

For now the default remains SW_HIDE to avoid introducing any
unintentionally breaking changes to the workflows of existing users, but
the sample configuration has been updated to set SW_MINIMIZE given that
it supports alt-tabbing across workspaces and through all levels of all
window containers, which seems to me to be more new user-friendly
behaviour.

re #72